relocate transactions table
