+++ packages/client/src/components/charges/charge-transactions-table.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 import { ChargeTableTransactionsFieldsFragmentDoc } from '../../gql/graphql.js';
 import { FragmentType, getFragmentData } from '../../gql/index.js';
-import { TransactionsTable } from '../common/index.js';
+import { TransactionsTable } from '../transactions-table/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
 /* GraphQL */ `
