+++ packages/client/src/components/charges/extended-info/creditcard-transactions-info.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 import { CreditcardBankChargeInfoFragmentDoc } from '../../../gql/graphql.js';
 import { FragmentType, getFragmentData } from '../../../gql/index.js';
-import { TransactionsTable } from '../../common/index.js';
+import { TransactionsTable } from '../../transactions-table/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
 /* GraphQL */ `
