+++ packages/client/src/components/common/index.ts
@@ -17,4 +17,3 @@ export * from './merge-businesses/index.js';
 export * from './modals/index.js';
 export * from './simple-grid.js';
 export * from './tooltip.js';
-export * from './transactions-table/index.js';
