--- packages/client/src/components/common/transactions-table/cells/index.ts
@@ -1,7 +1,0 @@
-export { DebitDate } from './debit-date.js';
-export { EventDate } from './event-date.js';
-export { Amount } from './amount.js';
-export { Account } from './account.js';
-export { Description } from './description.js';
-export { SourceID } from './source-id.js';
-export { Counterparty } from './counterparty.js';
