+++ packages/client/src/components/transactions-table/cells/amount.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
-import { TransactionsTableAmountFieldsFragmentDoc } from '../../../../gql/graphql.js';
-import { FragmentType, getFragmentData } from '../../../../gql/index.js';
-import { formatStringifyAmount } from '../../../../helpers/index.js';
+import { TransactionsTableAmountFieldsFragmentDoc } from '../../../gql/graphql.js';
+import { FragmentType, getFragmentData } from '../../../gql/index.js';
+import { formatStringifyAmount } from '../../../helpers/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
 /* GraphQL */ `
