+++ packages/client/src/components/transactions-table/cells/debit-date.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 import { format } from 'date-fns';
-import { TransactionsTableDebitDateFieldsFragmentDoc } from '../../../../gql/graphql.js';
-import { FragmentType, getFragmentData } from '../../../../gql/index.js';
+import { TransactionsTableDebitDateFieldsFragmentDoc } from '../../../gql/graphql.js';
+import { FragmentType, getFragmentData } from '../../../gql/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
 /* GraphQL */ `
