+++ packages/client/src/components/transactions-table/cells/description.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from 'react';
-import { TransactionsTableDescriptionFieldsFragmentDoc } from '../../../../gql/graphql.js';
-import { FragmentType, getFragmentData } from '../../../../gql/index.js';
+import { TransactionsTableDescriptionFieldsFragmentDoc } from '../../../gql/graphql.js';
+import { FragmentType, getFragmentData } from '../../../gql/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
 /* GraphQL */ `
