+++ packages/client/src/components/transactions-table/cells/event-date.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 import { format } from 'date-fns';
-import { TransactionsTableEventDateFieldsFragmentDoc } from '../../../../gql/graphql.js';
-import { FragmentType, getFragmentData } from '../../../../gql/index.js';
+import { TransactionsTableEventDateFieldsFragmentDoc } from '../../../gql/graphql.js';
+import { FragmentType, getFragmentData } from '../../../gql/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
 /* GraphQL */ `
