+++ packages/client/src/components/transactions-table/cells/index.ts
@@ -1,0 +1,7 @@
+export { DebitDate } from './debit-date.jsx';
+export { EventDate } from './event-date.jsx';
+export { Amount } from './amount.jsx';
+export { Account } from './account.jsx';
+export { Description } from './description.jsx';
+export { SourceID } from './source-id.jsx';
+export { Counterparty } from './counterparty.jsx';
